Split MCP connection manager module
Move McpConnectionManager and its methods into manager.rs while keeping the remaining MCP connection support code in mcp_connection.rs. Update internal imports and tests to match the new module boundary without changing behavior.
