+++ codex-rs/codex-mcp/src/mcp_connection_manager_tests.rs
@@ -1,11 +1,34 @@
 use super::*;
+use crate::declared_openai_file_input_param_names;
+use crate::mcp_connection::CODEX_APPS_TOOLS_CACHE_SCHEMA_VERSION;
+use crate::mcp_connection::CodexAppsToolsCacheContext;
+use crate::mcp_connection::ElicitationRequestManager;
+use crate::mcp_connection::ManagedClient;
+use crate::mcp_connection::StartupOutcomeError;
+use crate::mcp_connection::ToolFilter;
+use crate::mcp_connection::ToolInfo;
+use crate::mcp_connection::elicitation_capability_for_server;
+use crate::mcp_connection::elicitation_is_rejected_by_policy;
+use crate::mcp_connection::filter_tools;
+use crate::mcp_connection::load_startup_cached_codex_apps_tools_snapshot;
+use crate::mcp_connection::read_cached_codex_apps_tools;
+use crate::mcp_connection::tool_with_model_visible_input_schema;
+use crate::mcp_connection::transport_origin;
+use crate::mcp_connection::write_cached_codex_apps_tools;
+use codex_config::Constrained;
 use codex_protocol::ToolName;
 use codex_protocol::protocol::GranularApprovalConfig;
 use codex_protocol::protocol::McpAuthStatus;
+use futures::FutureExt;
 use pretty_assertions::assert_eq;
+use rmcp::model::CreateElicitationRequestParams;
+use rmcp::model::ElicitationAction;
+use rmcp::model::ElicitationCapability;
+use rmcp::model::FormElicitationCapability;
 use rmcp::model::JsonObject;
 use rmcp::model::Meta;
 use rmcp::model::NumberOrString;
+use rmcp::model::Tool;
 use std::collections::HashSet;
 use std::sync::Arc;
 use tempfile::tempdir;
